refactor improve loader locality
