+++ src/components/Scrycard/CardTypes/Loading/index.tsx
@@ -1,4 +1,4 @@
-import Loader from "../../../Throbber";
+import Loader from "./Throbber";
 
 export default function LoadingCard() {
     return <Loader />;
